[kyria] Update keymap based off of Clee's
Use
https://github.com/clee/qmk_firmware/blob/eeafd55c30b4213fdc5ad6e2d9c57f7fb29fd41b/keyboards/crkbd/keymaps/clee/keymap.c
as inspiration for updating my layers.

I'm still not happy th the placement of space and backspace and might
move my RAISE key over or as a layertap for something, but I'm starting
to get the hang of this. Still crap at programming stuff with all the
symbols, and I really need to fix the mod-tap settings.